[BACKPORT 2025.1][#26621] YSQL: Compute and use appropriate query id in plan hints
Summary:
When looking up hints in the hint table, the variable that records whether or not
we have tried to retrieve hints for a query can get incorrectly reset when processing
the query against the hint table. The fix is to save and restore this variable's value
when querying the hint table. The new hint table cache avoids this bug because it
does not (recursively) run a query against the hint table. The new tests added in
this diff make sure query ids are stable across hint usage and different GUC
settings, and across hint table cache usage.
Jira: DB-15999

Original commit: 832b9c40572cef6927ef3d88fac50b3a69325012 / D44376

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestYbQueryId' (new test)
./yb_build.sh --java-test 'org.yb.pgsql.TestYbAsh'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPgStatStatements'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressExtension'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgHintPlan'

Reviewers: mihnea, gkukreja, jason, mtakahara

Reviewed By: mtakahara

Subscribers: mtakahara, yql, jason

Differential Revision: https://phorge.dev.yugabyte.com/D46048