[BACKPORT 2025.1][#27973] DocDB: Retry txn status request if necessary instead of failing requestors with read time > time of status
Summary:
**Original Motivation for the problem**
When DDLs were changed to acquire `SELECT FOR UPDATE` on catalog version at the start of the ddl, sequential DDLs on a single PG connection were facing conflict errors occasionally as follows:
(1) DDL1 commits but intents are not yet flushed
(2) DDL2 checks intents db for conflicts and finds conflicting intents from DDL1. If it is able to resolve the transaction status of DDL1 as committed, there is no conflict. However, sometimes, the transaction status is still found as pending, resulting in a conflict error.

This was the case, despite DDL1's commit time < DDL2's read time.

**Changes involved in this revision**
1. We batch transaction status requests at `running_transaction.cc`. This implies that if an active txn status request is in progress for `txn X`, we don't launch another one if a different txn requests status of `X`, but just queue it onto the existing set of waiters waiting for the status of `X` i.e. `status_waiters_`.

And when the txn status arrives, we seem to even extract waiters with `read time > time of status`, and then fail them with `TryAgain` and pg error code set to `kConflict`.

This is not necessary, since we don't know that txn status of X as of the new read time, and it could lead to unnecessary retries/failures.

This revision addresses the issue by not extracting waiters with `read time > time of status`, and automatically re-launching another status request hoping to see appropriate status with time of status >= new read time.

One other change is to remove the below DFATAL log as it is now possible for a waiter with higher serial number to be resumed when read time < time of status < global time.
```
      LOG_IF_WITH_PREFIX(DFATAL, waiter.serial_no > serial_no)
          << "Notify waiter with request id greater than id of status request: "
          << waiter.serial_no << " vs " << serial_no;
```

2. Transaction status requests launched as part of core conflict resolution `ConflictResolver::FetchTransactionStatuses` were setting `StatusRequest::serial_no` as 0. With this setting, the requestor's callback would always be invoked despite the relation between read time vs time of status as `it->serial_no <= serial_no` would hold true in `RunningTransaction::ExtractFinishedStatusWaitersUnlocked`. So the change in 1 alone wouldn't suffice for automatic txn status retries, and hence the revision also addresses this issue by setting the correct `serial_no` as part of status requests.

Once the request is done with core conflict resolution, the `WriteQuery` might perform some reads as part of `WriteQuery::DoCompleteExecute` which execute reads as of `WriteQuery::read_time_`, and could launch additional txn status requests. We now use the same `serial_no` for txn status requests launched as part of these reads as well (which was the original motivation for `serial_no`).

This is done by moving `RequestScope` from `ConflictResolver` to `WriteQuery`, and explicitly releasing it after execution of the core conflict resolution callback i.e. in case of pgsql write - after submitting the write for replication, and in case if cql - after submitting all `UpdateQLIndexesTask` tasks.

This shouldn't lead to any semantic/behavioral change since the original order of operations was
```
TabletServer forms write query -> core conflict resolution -> WriteQuery::DoCompleteExecute + (submit write replication/UpdateQLIndexesTask) -> ~ConflictResolver -> ... (after write is replicated) -> client callback -> ~WriteQuery
```
Earlier, the request scope was being released as part `~ConflictResolver`, and now the request scope is released explicitly after `WriteQuery::ExecuteDone`. We are not deferring the release of request scope to ~WriteQuery because of this potential [[ https://github.com/yugabyte/yugabyte-db/commit/c0f7fa068e1881e3f7e72fd40315b73bcd6150d5 | issue ]]

**User facing impact**
The conflict errors which were spurious and **weren't** because of the user workload will be eliminated due to this revision.

Jira: DB-17585

Original commit: a6ba56d5673415a8f7365a967be819553d442599 / D45376

Test Plan:
Jenkins

Testing details included in JIRA. Ran auto analyze tests with DDL serialization which confirmed elimination of spurious conflicts being observed earlier.

Reviewers: sergei, pjain, patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D45980