[#28261] YSQL: support tserver cache bounded staleness for auth pg backend
Summary:
When --ysql_enable_read_request_cache_for_connection_auth=true, we use tserver
cache for pg auth backend's auth processing. In addition, to avoid the RPC
associated with fetching the latest catalog version from master, we use shared
memory catalog version as an approximation for the latest master catalog
version. The shared memory catalog version is one heartbeat interval away from
the latest master catalog version. This means that in normal situations, it is
only 1-second behind because the heartbeat interval is 1 second. However in
abnormal situations, we may not have a successful heartbeat due to network
flakiness or other reasons and the delay is currently unbounded. It is good to
provide a guarantee of bounded staleness that can be configured via a gflag.

There is already a tserver cache expiration mechanism built in. This diff adds a
new gflag --pg_cache_response_trust_auth_lifetime_limit_ms that causes tserver
cache entry for pg auth backend to expire after 60 seconds (the default value).
In order to separate tserver cache entries that are used for regular backends, I
added a new prefetch mode PrefetchingCacheMode::TRUST_CACHE_AUTH so that the new
gflag only applies to pg auth backends.

Note that the way tserver cache works is that when there are multiple concurrent
requests to the same cache entry that is expired, all of the requests will be
blocked until the tserver completes rebuilding the cache entry by sending RPC to
master.

In addition, even though the shared memory catalog version is stale, the cache
entry will contain the latest data from master for the latest master catalog
version but it will be tagged with the stale shared memory catalog version.

For example, if we have lost heartbeat and shared memory catalog version is X,
latest master catalog version is Y (X < Y), then after the cache entry expires
we'll rebuild the tserver cache entry for version X with the latest catalog data
of version Y. That is the version X tserver cache entry is filled with the
latest catalog data of version Y. Later when heartbeat resumes and the shared
memory catalog version is updated to Y, it will trigger another cache entry
build which ends up having exactly the same cache data but now tagged with
version Y.

A new unit test is added to demonstrate that an old password can still be used
and new password does not take effect before
--pg_cache_response_trust_auth_lifetime_limit_ms is reached. This new unit test
would fail before this change because heartbeat cache refresh is disabled so the
old password never expires.
Jira: DB-17931

Test Plan: ./yb_build.sh release --cxx-test pgwrapper_pg_catalog_perf-test

Reviewers: kfranz, sanketh

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46050