[#28124] xCluster: backfill namespace_id before using it in AllowedHistoryCutoff
Summary:
For #16551 we added to TSTabletManager::AllowedHistoryCutoff:
```
auto xcluster_safe_time_result =
      server_->GetXClusterContext().GetSafeTime(metadata->namespace_id());
```

This however assumes that metadata->namespace_id() has been loaded by a previously called TabletPeer::GetNamespaceId().
This is not guaranteed, and could lead to issues with not properly using the xcluster safe time for the history cutoff.

Eg, this is seen with replicated_ddls table for automatic mode replication, as it gets cleaned up by compactions despite being a part of an xcluster namespace.

In this diff, we fix this by backfilling if needed immediately before we use the namespace_id in TSTabletManager::AllowedHistoryCutoff.

Pulled out the backfilling code from TabletPeer::GetNamespaceId() into a static function so that TSTabletManager::AllowedHistoryCutoff can also use it.  Added some comments warning about the need for backfilling for other users of this field.

Did some minor tidying.

Fixes #28124
Jira: DB-17767

Test Plan:
Modified the following test to check that the namespace_id is filled in after TSTabletManager::AllowedHistoryCutoff is called:
```
ybd release --cxx-test xcluster_safe_time-itest --gtest_filter '*.ConsumerHistoryCutoff'
```
Verified that this test fails if we do not do the new backfilling of this diff.

Reviewers: xCluster, hsunder

Reviewed By: hsunder

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46075