[#28206][#28218][#28231][#28242] YSQL: Rewrite Cascading Timeout fix to affect only auth
Summary:
This diff rewrites the fix introduced in D45580 for cascading timeouts. The earlier fix applied the
change inside `od_router_attach()` which affected even transaction backend attach calls in addition
to calls during auth. While the new codepath was handled correctly in the auth_backend section,
other calls to attach weren't taken care of in places like `od_frontend_attach()` (and even
passthrough was missed out on in the fix :/).

This caused an implicit invariant to break - odyssey assumed that if `client->server` is non `NULL`,
then the server is successfully attached and server->io.io is non-null. This wasn't the case with
the new fix and so deallocating server->io.io would lead to a segmentation fault.

This new fix introduces handling for cascading timeouts _outside_ the call to `od_router_attach()` in
both `yb_auth_via_auth_backend()` and `yb_execute_on_control_connection()`.

For a comprehensive explanation of the issue, refer to Jira ticket [[ https://yugabyte.atlassian.net/browse/DB-17906 | DB-17906 ]]

A junit test is to be added for this in this revision itself.
Jira: DB-17865, DB-17875, DB-17894, DB-17906

Test Plan: Jenkins: urgent, all tests

Reviewers: skumar, asrinivasan, mkumar, arpit.saxena, #db-approvers

Reviewed By: skumar, arpit.saxena, #db-approvers

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D45994