[BACKPORT 2024.2.5][#28014] CDC: Do not update explicit checkpoint if tablet safe time is higher than slot restart time
Summary:
**Backport description:**
1. Merge conflict in cdcsdk_virtual_wal.h where some other changes were backported other than the private field last_persisted_record_id_commit_time_ and the method ShouldPopulateExplicitCheckpoint().
2. Merge conflict in TestPgReplicationSlot where some other tests got backported too.

In `UpdateAndPersistLSN` we first trim the maps and then update the slot entry in `cdc_state` table. If the write to cdc state table fails, then we end up with trimmed maps but the persisted data is left behind. Now in the next `GetChanges` call we will send explicit checkpoint on the basis of the trimmed checkpoint. Hence, the safe hybrid time for individual tablets could be greater than the persisted slot restart time.

Now if there is a restart, walsender will fetch schema as of the persisted restart time. But the first record sent by VWAL will have commit time higher than the restart time. If there was a DDL (or any other op) with commit time between the restart time and the safe hybrid time, then it will be missed here. As a result walsender will fail to decode the records sent after restart.

Consider the following scenario:
1. We have the slot restart time as 100 and a single tablet whose safe hybrid time is 100 as well.
2. We now have a DDL at time 150 and another DML at time 160.
3. Once we call the `UpdateAndPersistLSN`, the internal map of checkpoints will be truncated till 160.
4. Now the update to the `cdc_state` table fails, so will end up with a slot restart time of 100 and tablet safe time value of 160.
5. In case there is a restart now, we will rebuild VWAL and the schema built here would be built as of slot restart time i.e. 100 but the first record sent by VWAL would be the one with commit time 160 i.e. having a commit time higher than slot restart time.
6. As a result, walsender will not be able to decode those events sent after the restart and hence the data loss.

As a fix to prevent this from happening, this diff introduces a logic where we do not send explicit checkpoint to the individual tablets if their safe hybrid time exceeds the slot restart time.

Additionally, this diff also modifies the type for the `safe_hybrid_time` to `uint64_t` from the correctness perspective.
Jira: DB-17632

Original commit: 3b536e8f20e21bb083b940ff75561203fcecb4f2 / D45779

Test Plan:
./yb_build.sh --cxx-test cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestCheckpointNoUpdateUponCdcStateUpdateFailure
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testNoFailureWhenCdcStateUpdateFails'

Reviewers: skumar, asrinivasan, sumukh.phalgaonkar

Reviewed By: sumukh.phalgaonkar

Subscribers: ycdcxcluster, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D46059