[PLAT-18203] Ignore machine_image for edit universe flow
Summary:
In our code machineImage still take precedence over imageBundleUUID. That leads to a fact that if full move happens, new nodes will still be using old machine image.
That means that there could be nodes that have a discrepancy between imageBundleUUID and actual image installed.
To get rid of machineImage we can start with ignoring that field in edit flow.
After all the clusters that have that discrepancy are fixed we can remove that reference completely.

Test Plan:
Set incorrect machineImage for existing universe
Do full move - verify it is succesful

Reviewers: svarshney, amindrov

Reviewed By: amindrov

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D45787