[#28201] YSQL: Auto analyze pre-emption logic update
Summary:
A DDL statement like

`create table fk(id int primary key references pk(id));`

can still be preempted by an auto-ANALYZE even with the logic in 9e1c574 because the lock on catalog version row (in pg_yb_catalog_version) is only taken if we are at nesting level 0 AND we know that the nesting level 0 DDL work will be doing a catalog version increment.. For the fk table create, PG executes the DDL in two parts. A create table (at YB nesting level 0) and the alter table fk reference addition (at YB nesting level 1).

While taking a lock at nesting level 1 would fix this,  it introduces a potential for ANALYZE to commit in between the two nesting levels and preempt the user DDL, requiring a further enhancement to perform a full statement level retry of the DDL.

9e1c574 avoided taking a `for update` lock on DDLs that do not increment catalog versions so that concurrent `CREATE TABLE`   could still execute.

Instead, this diff changes the pre-emption logic to be the following:
1. Regular DDL takes a lock on the catalog version row `FOR KEY SHARE` with a high priority.
2. ANALYZE DDL triggered by auto analyze takes a lock on the catalog version row `FOR UPDATE` with a low priority.

With this change, ANALYZE triggered by auto analyze is always aborted by concurrent DDL. Concurrent non-ANAYZE DDL do not abort each other at the start of the DDL (they still conflict on the catalog version row at the end as always), so we lose that feature of commit 9e1c574, but that is not a strict requirement for auto-analyze to be turned on by default.
Jira: DB-17853

Test Plan: A new test verifies the different cases involved.

Reviewers: pjain, yguan

Reviewed By: pjain, yguan

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D45870