[#27575] DocDB: Take master snapshot metadata as of the snapshot hybrid time
Summary:
As part of creating a snapshot for a database, we take a snapshot of the master metadata related to the backup and save it in `SysSnapshotEntryPB`, which is persisted to disk. This snapshot entry contains, in addition to other things, the following:
- The sys catalog entries for the namespaces, tables, and tablets that are included in the snapshot.
- snapshot_hybrid_time: The snapshotted tablets are consistent as of this hybrid time. i.e, it is the timestamp at which all data written after it will not be visible at restore, and all data written before this timestamp must exist in the snapshot.

**Issue with selecting snapshot_hybrid_time**
The `snapshot_hybrid_time` is picked up at the master leader after receiving the `CreateSnapshot` RPC and is calculated as: `Clock()->MaxGlobalNow()`
`MaxGlobalNow` returns the maximum time that can be found in all the cluster nodes. I.e, it is equal to `physical_now + max_clock_skew_usec`. The default value of `max_clock_skew_usec` is 500 ms.
We choose this as the snapshot hybrid time so that the read-after-write semantics are respected in snapshots. i.e, If a user writes some data and then creates a snapshot, we must ensure that this data is included in the snapshot.

We can see that the snapshot hybrid time is selected to be a little bit in the future. This can cause backup inconsistency between the ysql_dump metadata (as ysql_dump is taken as of `snapshot_hybrid_time`) and the snapshot metadata that is collected at the master side (which has metadata earlier than `snapshot_hybrid_time`). To illustrate the issue, consider the following example where we have a simple table `my_table(k int, v int)`:

1- `CreateSnapshot` RPC received by master at timestamp t1.
2- Master leader collects `SysTableEntryPB` for all the tables, including `my_table` with DocDB schema `(k int, v int)`
3- Master leader marks `snapshot_hybrid_time` to `t3 = MaxGlobalNow() = physical_now + max_clock_skew_usec`. Notice that t3 might be slightly in the future.
4- An `alter table my_table add column v2 int` comes from the client and commits at time `t2` < `t3`.
5- Run ysql_dump as of `snapshot_hybrid_time = t3`.

We can see that ysql_dump has the new schema of `my_table`. i.e. `(k int, v int, v2 int)`. However, the snapshot has the old schema of `my_table` as per step 2.

The source of the issue is that the snapshot entries persisted at the master side are as of timestamp t1, however, the snapshot_hybreid_time is t3.

The diff fixes the issue by taking a snapshot of the metadata that is consistent as of `snapshot_hybrid_time`. This is done as follows:
1- `CreateSnapshot` RPC received by master at timestamp t1.
2- The master leader selects the `snapshot_hybrid_time = t2 = MaxGlobalNow()` as usual.
3- The master leader waits until the `snapshot_hybrid_time` is a safe time. That guarantees that no new writes will come to the sys.catalog tablet with `commit_hybrid_time` earlier than t2.
4- The master leader reads the sys.catalog tablet "from disk" and collects the snapshot entries (namespaces, tables, and tablets) as of snapshot_hybrid_time.
5- The master continues the snapshot creation workflow on the tservers as usual.

This guarantees that both ysql_dump output and SnapshotInfoPB are consistent and are as of `snapshot_hybrid_time`.

**When to opt for the new workflow vs old workflow of creating the snapshot**
The new workflow will only be used if the client is providing only the `namespace_id` to the master leader in the `CreateSnapshotRequestPB`. Otherwise, the old workflow of collecting the entries from the in-memory state is used.

The old workflow (where the backup fails in the event of concurrent DDL): YBC selects the set of DocDB tables to be included in the snapshot (using the `ListTables` RPC) and passes this list to the master as part of the `CreateSnapshot` RPC.

The new workflow (where backup during DDL is supported): YBC only sends a `CreateSnapshot` RPC to the master leader with the `namespace_id` only. This means that the master collects the set of tables that are committed as of `snapshot_hybrid_time` and belongs to namespace `namespace_id`.
Jira: DB-17138

Test Plan: ./yb_build.sh --cxx-test yb-backup-during-ddl-test --gtest_filter Colocation/YBBackupTestWithColocationParam.CreateConsistentMasterSnapshot/0

Reviewers: asrivastava, zdrudi

Reviewed By: asrivastava, zdrudi

Subscribers: mihnea, yguan, ybase, hsunder

Differential Revision: https://phorge.dev.yugabyte.com/D46081