[BACKPORT 2025.1][#23908] docdb: Add/fix metrics for cluster balancing time/data estimates
Summary:
This diff improves the existing metric for how far a cluster is from balanced (`total_table_load_difference`) and adds a new `total_estimated_data_to_balance_bytes` metric that reports the an estimate of the amount of data that must be copied (remote bootstrapped) in the cluster for the cluster to be balanced. This makes it easier for users to estimate how long cluster balancing will take.

The improved metrics also act as a check that the cluster balancing algorithm is correct. If the metric is not 0 but the cluster balancer is idle, the algorithm has not found the optimal tablet distribution.

**Context**
The `total_table_load_difference` metric used to report the sum across tables of the **table load difference**, which was the number of tablets on the most loaded tserver (for a given table) minus the number of tablets on the most loaded tserver (for the same table). This metric had some issues:
- It was not 0 in some balanced clusters. Specifically, it did not handle custom placement info or clusters with different numbers of tservers in each zone, both of which could result in more tablets on some tservers than others.
- It did not help users predict how long cluster balancing would take because:
  - It did not account for a table's data size
  - The metric went down slower if there are many tservers: the table load difference would be 2 if a cluster's most loaded tserver for a table had 4 tablets and the least loaded had 2 tablets. But, the metric would also be 2 if there were 100 most loaded tservers with 4 tablets and all the least loaded tservers had 2 tablets.

**New calculation for total_table_load_difference**
This diff adds a function, `CalculateOptimalLoadDistribution` (O(num_tservers + num_placement_blocks)), which explicitly calculates the closest optimal tablet distribution to the current tablet distribution (fewest moves) for a given table. Another function, `CalculateTableLoadDifference`, calculates the new table load difference as the number of replicas that need to be added (remote bootstrapped) to reach the optimal distribution (we ignore removes because those should be relatively fast).

`total_table_load_difference` is now the sum of the number of the table load difference (as given above) across all tables. This is only a lower bound on the number of moves the cluster balancer must make because it does not account for global load balancing (balancing across tables). This should still be fairly accurate because we break ties when moving tablets by global load, so we should never have many global load balancing moves to perform.

**total_estimated_data_to_balance_bytes**
`total_table_load_difference` does not account for the size of each table and thus cannot be used to infer how long cluster balancing would take. To address this, this diff adds a `total_estimated_data_to_balance_bytes` metric, which is defined as the sum across all tables of the table load difference multiplied by the average size of a tablet for that table. This might be skewed for non-uniform tables. If we further assume that cluster balancing is bound by the data transfer time (though this is not quite the case until all cluster balancer performance improvements land), a user could get a time estimate by dividing this metric by `FLAGS_remote_bootstrap_bytes_per_sec`.

I ran the tests with a LOG(FATAL) in `ClusterLoadBalancer::IsIdle` to check whether the metric is 0 when the cluster balancer says it is idle. It usually is, but there are cases where the cluster balancer does not converge on the optimal distribution (e.g., GitHub issue #15853: moving tablets across zones).
Jira: DB-12813

Original commit: f5ac1b5093c8e08bec6b95783648f59a0dd63623 / D45002

Test Plan:
`./yb_build.sh --cxx-test master_load_balancer_mocked-test --gtest_filter OptimalLoadDistributionTest.*`

I also modified the following test:
`./yb_build.sh --cxx-test load_balancer_mini_cluster-test --gtest_filter "LoadBalancerMiniClusterTest.TableLoadDifferenceMetric"`

I also tested both metrics when scaling out a cluster from 3 -> 4 nodes with 2 tables of 100 tablets each (with ~50 MB and ~130 MB of data). It's easier to read the amount of time until the cluster balances from the estimated bytes metric.
{F373421}
{F373420}

Reviewers: zdrudi, mlillibridge

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46091