[BACKPORT 2025.1][#28270] YSQL: Propagate additional locality information from PG to TServer
Summary:
From the ticket:

We currently propagate the return value of YBCIsRegionLocal(relation) through Perform
in order to determine whether or not a transaction should be started as region-local or global.
This actually checks multiple conditions:

* that we are in normal processing mode
* that the relation is not a system relation
* that the tablespace cost is at most the interzone cost (effectively: that the tablespace is in the same region as the node)

In preparation for adding categorization of transactions into new tablespace-local localities, we
will need to know the actual tablespace oid itself (since there will be different a tablespace-local
locality for each tablespace). Additionally, we still care about the first two conditions
(normal processing mode/not system relation) and need to separate that check from the
is-same-region check.

This diff adds tablespace oid (optional uint32 tablespace_oid) in each `PgsqlReadRequestPB` and
`PgsqlWriteRequestPB` if the table read/write has a user defined tablespace. Because we already set
catalog version in `PgsqlReadRequestPB` and `PgsqlWriteRequestPB`, I added a call to the new
function `YbMaybeSetNonSystemTablespaceOid` whenever `YbSetCatalogCacheVersion` is called. This
ensures whenever catalog version is set, we also get a chance to set tablespace oid if needed.

There are two scenarios that we do not set catalog version in a read/write request.
(1) when internally read/write a system relation
(2) in foreign key cache, we issue read/write to prefetch the latest FK values.

For (1), we are fine because a system relation will not have a user-defined tablespace.
For (2), I build a table_id to tablespace_oid map `tablespace_map_`. Each time `RelationCacheInsert`
is called to enter a relation into the relcache, the new function `YBCRecordTablespaceOid` is called
to record the pair (table_id, tablespace_oid). Each time `RelationCacheDelete` is called to delete a
relation from the relcache, the new function `YBCClearTablespaceOid` is called to remove the key
`table_id` from the relcache.

The new map `tablespace_map_` is passed down to FK cache so that it can be used to build up the
read/write request if the `table_id` has an entry in the map.

In debug build, I also added `YBCPgCheckTablespaceOid` to ensure when
`YbMaybeSetNonSystemTablespaceOid` is called and does set a user defined tablespace oid, we
will find a matching entry in the `tablespace_map_` that has the same tablespace oid.

**Upgrade/Rollback safety:**

The new field `tablespace_oid` in `PgsqlWriteRequestPB` and `PgsqlReadRequestPB` is
optional and will be ignored by an old tserver that does not have it. Its
purpose is for better performance by taking into account tablespace information.
If a new tserver receives a `PgsqlWriteRequestPB` or `PgsqlReadRequestPB`
from an old tserver, the new field `tablespace_oid` will not be set and the new
tserver just won't do the performance optimization either.
Jira: DB-17945

Original commit: 22c814e6295cfdc587a3d7b17c9c837d6215f549 / D46099

Test Plan:
jenkins
Manually tested in the context of https://phorge.dev.yugabyte.com/D46211.

Reviewers: kfranz, sanketh

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46807