[#27621] docdb: Block clone while table is altering
Summary:
`GetBackupEntriesAsOfTime` currently filters out tables that are not in the RUNNING state as of the read time. If a table is altering, we need to read the PG catalog as of the clone time to figure out which table is committed. Currently, we clone the table without any data and this caused silent data loss. After this diff, we fail the clone.

We should eventually support cloning to the time when a DDL is occurring. This is tracked by https://github.com/yugabyte/yugabyte-db/issues/28814.
Jira: DB-17211

Test Plan: `./yb_build.sh release --cxx-test tools_yb-admin-snapshot-schedule-test --gtest_filter=YbAdminSnapshotScheduleTestWithYsqlColocationRestoreParam.RestoreWithBackfillingIndex/NotColocated_Clone -n 10 `

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46130