[PLAT-18121]: Use pod config instead of AZ config while executing commands on DB pods.
Summary:
The PG upgrade t-server check fails to find the arch of DB pods, as in order to find the arch, YBA needs to connect to DB pods and exec command, but if the K8s universe is deployed using multiple service accounts, then it fails with the error "insufficient permission," as during the task, it uses the YBA service account instead of the DB pods service account.
We will now use Kube-config per pod instead of AZ-level override config, which will merge AZ and provider-level overrides.

Test Plan:
NA
Will coordinate with @yash.priyam to get it tested manually.

Reviewers: vkumar, anijhawan

Reviewed By: vkumar, anijhawan

Subscribers: yash.priyam, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46182