[#28069] YSQL: Simplify PgTableCache interface
Summary:
Current signature of the `PgTableCache::GetTables` is over-complicated. The caller code requires lots of boilerplate code because `PgTableCache::GetTables` accepts two independent shared pointer for result and for listener that result is ready. This implies that result and a listener may have different lifetime, but in real usage both has same lifetime.

Example of boiler plate caller's code

```
   auto listener_and_result = std::make_shared<
        std::pair<SharedMemoryPerformListener, PgTablesQueryResult>>();
    auto listener = SharedField(listener_and_result, &listener_and_result->first);
    auto result = SharedField(listener_and_result, &listener_and_result->second);
    table_cache().GetTables(table_ids, {}, result, listener);
```

Also current signature requires to build/use result object (the collection of tables which uses mutex internally and some other additional stuff) even in trivial case when the only one table is required.

This diff simplifies the `PgTableCache::GetTables` signature. Instead of accepting two independent shared pointers it accepts the single one of type `PgTablesQueryTableReceiver`.
```
class PgTablesQueryTableReceiver {
 public:
  virtual void Ready(const Result<TableInfo>& info, size_t total_tables_count) = 0;
};
```

In contrast to current behavior where the `PgTablesQueryListener::Ready` method get called when all tables from the query is ready, the `PgTablesQueryTableReceiver::Ready` method is called for each table individually.

To achieve functionality provided by the removed `PgTablesQueryListener` interface the  `PgTablesQueryResultReceiver` helper abstract class is created. Its method `PgTablesQueryTableReceiver::TablesQueryResultReady` is called when all the tables are ready. Also this class provides access to the embedded `PgTablesQueryResult` object.

In addition this diff changes the `PgTablesQueryResult` class. Prior to the change this class has mixed up thread safe and non thread safe functions. There is the mutex which allows to fill the object with new tables asynchronously in thread safe manner but getter methods doesn't lock this mutex.
As a result it is not safe to call getters prior to the moment when object is fully ready. Such access is not intent but nevertheless this makes internals of the `PgTablesQueryResult` a little weird.

In current version access to the incompletely built `PgTablesQueryResult` is impossible and the class is fully thread safe. Also it contains getters methods only. All stuff related to asynchronous building is moved into the `PgTablesQueryResultReceiver` object.
Jira: DB-17700

Test Plan: jenkins

Reviewers: sergei, bkolagani, myang

Reviewed By: sergei

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45542