[#28331] docdb: Support multiple sets of tablets when checking tablet creation limits.
Summary:
We currently only support checking one replication info and number of tablets when checking tablet limits:

```
Status CanCreateTabletReplicas(
    int num_tablets, const ReplicationInfoPB& replication_info,
    const TSDescriptorVector& ts_descs, const BlacklistSet& blacklist) {
```
But this does not allow us to use the check when creating multiple tables at once (e.g., for DB
cloning). This diff adds an overload that takes a vector of replication infos and the number of
tablets to create for each.

Also adds a flag to globally control the number of tablets supported in a cluster.
Jira: DB-18013

Test Plan: Added some tests to `./yb_build.sh release --cxx-test tablet_creation_limits_test`

Reviewers: zdrudi

Reviewed By: zdrudi

Subscribers: mlillibridge, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46209