[BACKPORT 2025.1][#28125] YSQL: Defer txn priority assignment to first read/write for serializable isolation DMLs
Summary:
For DMLs, YSQL sets the transaction priority when `isolation_level_` is being set as part of `PgTxnManager::CalculateIsolation`, i.e the first op of the YSQL transaction where it makes a decision that it would need a corresponding docdb transaction, and hence explicitly sets the isolation level so that a docdb txn gets created while processing of the request at the `PgClientSession` layer.

With fail-on-conflict enabled, transactions can have different priorities. In case of conflicts, the lower priority transaction(s) are aborted. There is custom logic to assign txn priorities based on the statement type.

When object locking is enabled, a few of the fail-on-conflict tests using serializable isolation fail. This is because wrong transaction priorities are being set as part of the first object lock call whereas earlier they were being set as part of the first read/write op. For instance, consider the following
```
begin transaction isolation level serializable;
select * from test ... for update;
```
Without object locking - txn priority takes into account the row lock markers and we compute the priority to be something in the interval `[kHighPriTxnLowerBound, kHighPriTxnUpperBound]`
With object locking - txn priority wrongly gets sets to something in the interval `[kRegularTxnLowerBound, kRegularTxnUpperBound]` on the first object lock call, and few tests fails with unexpected serialization errors like the below
```
could not serialize access due to concurrent update
```

This revision addresses the issue for serializable isolation DMLs by deferring txn priority assignment to the first read/write op.

**Note**: The problem doesn't exist with RC/RR DMLs (without savepoints) since the first few object lock calls compute the effective isolation level to be `NON_TRANSACTIONAL`, until a read with row marker/write op inits `isolation_level_` and rightly sets the txn priority. But the problem exists for RR isolation transactions which issue a savepoint in the beginning, and then execute a read with a row marker. For instance,
```
begin transaction isolation level repeatable read;
savepoint a;
select * from t ... for update;
```
Since the select is done in 2 read ops, the first without a row marker followed by a read with a row marker, the first read op wrongly sets the priority in `PgTxnManager::CalculateIsolation` due to the below `if` clause.
```
  if (PREDICT_FALSE(enable_table_locking_) && active_sub_transaction_id_ > kMinSubTransactionId &&
      pg_isolation_level_ != PgIsolationLevel::READ_COMMITTED) {
    read_only_op = false;
  }
```

Not fixing this problem since transaction priorities don't come in play when what-on-conflict is enabled, which is true for releases > 2024.1
Jira: DB-17761

Original commit: 27a57fc4465fa54bfddc29f207b9ac1e81681e57 / D45671

Test Plan:
Enabled object locking feature and ran the below
```
./yb_build.sh --cxx-test='TEST_F_EX(PgLibPqTest, TxnConflictsForTablegroupsYbSeq, PgLibPqFailOnConflictTest) {' -n 10 --tp 1
./yb_build.sh --cxx-test pg_row_lock-test --gtest_filter *PartialKeyRowLockConflict*
```

Reviewers: pjain, sanketh, rthallam

Reviewed By: pjain

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46034