[#28333] YSQL: Integrate with PG's snapshot management for Repeatable Read isolation level
Summary:
edbd06b added support for integrating YSQL with PG's snapshot management which
allows switching snapshots. This was enabled exclusively for Read Committed
isolation level.

This change enables that support for Repeatable Read isolation as well. This is
a pre-requisite for fixing #26163 which relies on the fact that various foreign
key triggers use the latest snapshot instead of the transaction snapshot for
performing reads and hence require support for switching snapshots.
Jira: DB-18014

Test Plan: Jenkins

Reviewers: smishra, patnaik.balivada, dmitry, jason, stiwary, #db-approvers

Reviewed By: patnaik.balivada, dmitry, stiwary, #db-approvers

Subscribers: kramanathan, stiwary, svc_phabricator, aagrawal, jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43111