[BACKPORT 2025.1][PLAT-18358]: Fix PG-15 upgrade on cluster with DB audit logging enabled.
Summary:
During the PG-15 software upgrade task, we set the FlagInMemory subtask to update the YSQL major compatibility flag on the cluster before and after the upgrade. When forming the task, we used to pass the flag in an immutable object, which was incorrect, as we also update/set the audit logging flag. In order to add the audit logging-related flags, we should not override the flags object in the task params from an immutable to a mutable object inside the subtask, as if we ever pass an immutable object in task params, the subtask will handle it.
Original diff/commit: b4aafcd86937ffb902ccd28a8e8fc2354f33b8ba/D46221

Test Plan: Tested manually

Reviewers: anabaria, svarshney

Reviewed By: svarshney

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46288