[PLAT-18267] Default to always include inherited values when fetching runtime configs
Summary: Right now the client code defaults to false when not specified explicitly. This diff updates the default to true as a code search reveals we’re always manually passing true in all references to this function and this makes sense for the most common usage case: reading the current runtime config value at a particular scope.

Test Plan:
All other existing usage explicitly sets `includeInherited` to true:
{F385623}
only the usage in UniverseDisplayPanel is affected.

Test for UniverseDisplayPanel change.
Reset runtime config values to default.
Create a universe and remove node agent from one of the nodes.
Verify we're able to read the node agent scan interval runtime config and
that we show a warning banner as expected.
{F385622}

Reviewers: nsingh, rmadhavan

Reviewed By: nsingh

Differential Revision: https://phorge.dev.yugabyte.com/D46236