[PLAT-18346] Switchover task is failing while DDL replication is enabled
Summary:
The replicated_ddls table was made a user table in the recent versions of YBDB. The switchover precheck to ensure all the tables are in replication fails because the table replicated_ddls is only a metadata table which is not part of the replication group.

This diff removes it from the tables list so the precheck does not involve it.

Test Plan: - Created a DR config with automatic DDL and made sure after this change, the user is able to do a switchover

Reviewers: amindrov, spothuraju, jmak, nbhatia

Reviewed By: jmak

Differential Revision: https://phorge.dev.yugabyte.com/D46242