[#28355] docdb: Add webserver endpoint for creating perf flamegraphs
Summary:
This diff adds a new **beta** server endpoint: `/perf`.

**`/perf?seconds=1&freq=99`**
This endpoint runs `perf` on the system for the current process (tserver if using the 9000 port, master if using 7000) and generates a flamegraph using `flamegraph.pl` and `stackcollapse-perf.pl` (from https://github.com/brendangregg/FlameGraph, which is under the CDDL License). It also provides a link to download the collapsed stacks, which can be stored for later analysis or loaded into another stack viewer like speedscope.

**This endpoint is not exposed through any UI yet because we are still beta testing it.**

The endpoint accepts two URL parameters: `freq` and `seconds`. `freq` is the number of times to sample per second (99 by default), and seconds is how long to sample for (1s by default).

An example flamegraph is shown below. The stacks here are fragmented because we aggregate by the name of the stack, and some of our threads are named "TabletServer_0_Worker", "TabletServer_1_Worker", etc. and are thus not aggregated. This will be fixed in a future diff.
{F401695}
Jira: DB-18038

Test Plan: Ran a CQL workload and visited: `http://127.0.0.1:9000/perf?seconds=1&freq=999` to get the flamegraph in the summary.

Reviewers: kfranz, #db-approvers, hsunder

Reviewed By: kfranz, #db-approvers, hsunder

Subscribers: hsunder, svc_phabricator, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46264